requirements-test.txt
pytest>=8.0
pytest-timeout>=2.3
python-dotenv>=1.0
ldap3>=2.9
requests>=2.31
pyyaml>=6.0
pymysql>=1.1
pytest-playwright>=0.4.0
